Added rest error handling specific for exchange
Added ability to specify specific for exchange rest error handlers
