+++ core/src/exchanges/common.rs
@@ -254,7 +254,7 @@ impl ExchangeError {
             None,
         )
     }
-    pub(crate) fn unknown_error(message: &str) -> Self {
+    pub fn unknown_error(message: &str) -> Self {
         Self {
             error_type: ExchangeErrorType::Unknown,
             message: message.to_owned(),
