+++ exchanges/binance/Cargo.toml
@@ -24,6 +24,8 @@ serde_json = "1"
 sha2 = "0.9"
 tokio = { version = "1" }
 url = "2.0"
+hyper = { version = "0.14", features = ["http1", "runtime", "client", "tcp"] }
+function_name = "0.2.0"
 
 [dev-dependencies]
 actix-rt = "2"
