+++ exchanges/binance/src/lib.rs
@@ -16,4 +16,5 @@
 
 pub mod binance;
 pub mod exchange_client;
-pub mod support;
+
+mod support;
