+++ exchanges/binance/tests/binance/get_open_orders.rs
@@ -90,8 +90,7 @@ async fn get_open_orders_for_each_currency_pair_separately() {
 
     let exchange_account_id: ExchangeAccountId = "Binance_0".parse().expect("in test");
     let (api_key, secret_key) = get_binance_credentials_or_exit!();
-    let mut settings =
-        ExchangeSettings::new_short(exchange_account_id, api_key, secret_key, false, false);
+    let mut settings = ExchangeSettings::new_short(exchange_account_id, api_key, secret_key, false);
 
     settings.currency_pairs = Some(vec![
         CurrencyPairSetting::Ordinary {
