+++ exchanges/binance/tests/binance/wait_cancel_order.rs
@@ -16,8 +16,7 @@ async fn cancellation_waited_successfully() {
 
     let exchange_account_id: ExchangeAccountId = "Binance_0".parse().expect("in test");
     let (api_key, secret_key) = get_binance_credentials_or_exit!();
-    let mut settings =
-        ExchangeSettings::new_short(exchange_account_id, api_key, secret_key, false, false);
+    let mut settings = ExchangeSettings::new_short(exchange_account_id, api_key, secret_key, false);
 
     // Currency pair in settings are matter here because of need to check
     // Symbol in check_order_fills() inside wait_cancel_order()
