+++ exchanges/serum/Cargo.toml
@@ -31,6 +31,7 @@ solana-sdk = "1.7"
 spl-token = { version = "3.2", features = ["no-entrypoint"], default-features = false }
 tokio = { version = "1" }
 url = "2.0"
+function_name = "0.2.0"
 
 [dev-dependencies]
 core_tests = { path = "../../core_tests" }
